Achieving basic bot functionality.
I've made a bunch of tweaks, fixes, and several additions in this
commit.  I'm not going to list the fixes and most of the tweaks because
I don't remember a lot of them.  This is a big commit, at least for me.

Antiban has been added.
Several functions have been added to utils and used.
Bot now checks if target is in combat (has HP bar) before clicking.
The screen is now rotated between each attempt to use the stair, gate,
and bank.